+++ android/src/main/kotlin/com/example/inappstory_plugin/IASSingleStoryAdaptor.kt
@@ -6,9 +6,10 @@ import com.inappstory.sdk.externalapi.single.IASSingleStory
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 
 class IASSingleStoryAdaptor(
-        private val flutterPluginBinding: FlutterPlugin.FlutterPluginBinding,
-        private val appearanceManager: AppearanceManager,
-        private val iasSingleStory: IASSingleStory,
+    private val flutterPluginBinding: FlutterPlugin.FlutterPluginBinding,
+    private val appearanceManager: AppearanceManager,
+    private val iasSingleStory: IASSingleStory,
+    private val activityHolder: ActivityHolder,
 ) : IASSingleStoryHostApi {
     private val callback = IShowStoryOnceCallbackAdaptor(flutterPluginBinding)
 
@@ -18,22 +19,22 @@ class IASSingleStoryAdaptor(
         iasSingleStory.loadCallback(SingleLoadCallbackAdaptor(flutterPluginBinding))
     }
 
-    override fun showOnce(storyId: Long) {
+    override fun showOnce(storyId: String) {
         iasSingleStory.showOnce(
-                flutterPluginBinding.applicationContext,
-                storyId.toString(),
-                appearanceManager,
-                callback,
+            activityHolder.activity,
+            storyId,
+            appearanceManager,
+            callback,
         )
     }
 
-    override fun show(storyId: Long, slide: Long) {
+    override fun show(storyId: String) {
         iasSingleStory.show(
-                flutterPluginBinding.applicationContext,
-                storyId.toString(),
-                appearanceManager,
-                callback,
-                slide.toInt(),
+            activityHolder.activity,
+            storyId,
+            appearanceManager,
+            callback,
+            null,
         )
     }
 }