+++ example/lib/story_widget_simple_decorator.dart
@@ -72,6 +72,6 @@ class StoryWidgetSingleReader extends StoryWidgetSimpleDecorator {
 
   @override
   void onTap() {
-    IASSingleStoryHostApi().showOnce(storyId: story.id);
+    IASSingleStoryHostApi().showOnce(storyId: '${story.id}');
   }
 }